requirements.txt
django-countries==3.2
django-hvad==1.2.0
django-crispy-forms==1.4.0
django-vanilla-views==1.0.2
django-adminplus==0.3
django-toolbelt==0.0.1
dj-database-url==0.3.0
whitenoise==1.0.6
geojson==1.0.9
django-extensions==1.5.2
git+https://github.com/kennethreitz/tablib.git
git+https://github.com/aptivate/django-import-export.git
git+https://github.com/aptivate/django-admin-sortable.git
